### 2024-02-07 #6331 @editorialbot
[PRE REVIEW]: Soundata: Reproducible use of audio datasets
**Submitting author:** @magdalenafuentes (<a href="http://orcid.org/0000-0003-4506-6639">Magdalena Fuentes</a>)
**Repository:** https://github.com/soundata/soundata
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.3
