### 2024-02-07 #6331 @danielskatz
👋 @openjournals/dev - do you know why this command didn't run automatically?
