### 2024-02-07 #6331 @danielskatz
👋 @magdalenafuentes - thanks for your submission. Before we get started with the review, there are a couple of issues:

1. Your paper is quite a bit too long. JOSS has a normal upper limit of roughly 1000 words, and your paper is close to 3000.  Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for an example of a typical JOSS paper.

2. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.

Once you have made changes in the paper, whether this is content or references, please use the command `@editorialbot generate pdf` to make a new PDF.

editorialbot commands need to be the first entry in a new comment.

I'll put this on pause until you tell me it's ready to start, because these two issues have been resolved
