+++ paper.bib
@@ -2,14 +2,18 @@
   title={Datasets: A community library for natural language processing},
   author={Lhoest, Quentin and del Moral, Albert Villanova and Jernite, Yacine and Thakur, Abhishek and von Platen, Patrick and Patil, Suraj and Chaumond, Julien and Drame, Mariama and Plu, Julien and Tunstall, Lewis and others},
   journal={arXiv preprint arXiv:2109.02846},
-  year={2021}
+  year={2021},
+  doi = {10.48550/arXiv.2109.02846},
+  url = {https://doi.org/10.48550/arXiv.2109.02846}
 }
 
 @inproceedings{bittner2019mirdata,
   title={mirdata: Software for reproducible usage of datasets},
   author={Bittner, Rachel M and Fuentes, Magdalena and Rubinstein, David and Jansson, Andreas and Choi, Keunwoo and Kell, Thor},
   booktitle={ISMIR},
-  year={2019}
+  year={2019},
+  doi={10.5281/zenodo.3527750},
+  url={https://doi.org/10.5281/zenodo.3527750},
 }
 
 @article{pedregosa2011scikit,
@@ -19,7 +23,9 @@
   volume={12},
   number={Oct},
   pages={2825--2830},
-  year={2011}
+  year={2011},
+  doi={10.48550/arXiv.1201.0490},
+  url={https://doi.org/10.48550/arXiv.1201.0490}
 }
 
 @inproceedings{tensorflow,
@@ -28,7 +34,9 @@ author  = {Martin Abadi and Paul Barham and Jianmin Chen and Zhifeng Chen and An
 year  = {2016},
 URL = {https://www.usenix.org/system/files/conference/osdi16/osdi16-abadi.pdf},
 booktitle = {12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)},
-pages = {265--283}
+pages = {265--283},
+doi={10.48550/arXiv.1605.08695},
+url={https://doi.org/10.48550/arXiv.1605.08695}
 }
 
 @misc{tensorflow_datasets,
@@ -48,7 +56,9 @@ pages = {265--283}
   author={Zinemanas, Pablo and Hounie, Ignacio and Cancela, Pablo and Font Corbera, Frederic and Rocamora, Mart{\'\i}n and Serra, Xavier},
   booktitle={Ono N, Harada N, Kawaguchi Y, Mesaros A, Imoto K, Koizumi Y, Komatsu T, editors. Proceedings of the Fifth Workshop on Detection and Classification of Acoustic Scenes and Events (DCASE 2020); 2020 Nov 2-3; Tokyo, Japan.[Tokyo]: DCASE; 2020.},
   year={2020},
-  organization={Detection and Classication of Acoustic Scenes and Events (DCASE)}
+  organization={Detection and Classication of Acoustic Scenes and Events (DCASE)},
+  doi={10.5281/zenodo.4061782},
+  url={http://dx.doi.org/10.5281/zenodo.4061782}
 }
 
 @article{mesaros2016metrics,
@@ -59,7 +69,9 @@ pages = {265--283}
   number={6},
   pages={162},
   year={2016},
-  publisher={Multidisciplinary Digital Publishing Institute}
+  publisher={Multidisciplinary Digital Publishing Institute},
+  doi={10.3390/app6060162},
+  url={https://doi.org/10.3390/app6060162}
 }
 
 @misc{speechbrain,
@@ -69,16 +81,20 @@ pages = {265--283}
   eprint={2106.04624},
   archivePrefix={arXiv},
   primaryClass={eess.AS},
-  note={arXiv:2106.04624}
+  note={arXiv:2106.04624},
+  doi={10.48550/arXiv.2106.04624},
[…]
