### 2024-02-07 dbda5c6
Trim paper content
