+++ paper.md
@@ -50,7 +50,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-As research pipelines become increasingly complex, it is key that their different components are reproducible. In recent years, the machine listening research community has made considerable efforts towards standardization and reproducibility, by using common libraries for modeling and evaluation ([@tensorflow, @pedregosa2011scikit, @chollet2015keras, @mesaros2016metrics](#)), open sourcing models ([@zinemanas2020dcase, @speechbrain](#)) and data dissemination using resources such as [Zenodo](https://zenodo.org). However, it has been previously shown that discrepancies in the local version of the data and different practices in loading and parsing datasets can lead to considerable differences in performance results, which is misleading when comparing methods ([@bittner2019mirdata](#)). Besides, from a practical point of view, it is extremely inefficient to develop pipelines from scratch for loading and parsing a dataset for each researcher or team each time, and this increases the chances of bugs and differences that hinder reproducibility.
+As research pipelines become increasingly complex, it is key that their different components are reproducible. In recent years, the machine listening research community has made considerable efforts towards standardization and reproducibility, by using common libraries for modeling and evaluation ([@tensorflow, @pedregosa2011scikit, @mesaros2016metrics](#)), open sourcing models ([@zinemanas2020dcase, @speechbrain](#)) and data dissemination using resources such as [Zenodo](https://zenodo.org). However, it has been previously shown that discrepancies in the local version of the data and different practices in loading and parsing datasets can lead to considerable differences in performance results, which is misleading when comparing methods ([@bittner2019mirdata](#)). Besides, from a practical point of view, it is extremely inefficient to develop pipelines from scratch for loading and parsing a dataset for each researcher or team each time, and this increases the chances of bugs and differences that hinder reproducibility.
 
 
 There are other libraries that handle datasets, like `Tensorflow` ([@tensorflow](#)) or `DCASE-models` ([@zinemanas2020dcase](#)). However, using datasets in the context of those libraries makes it difficult to interchange models and software, as data loaders are designed to work with those environments and further adaptation is required to migrate them. Instead, we think that data should be handled separately, as a standalone library that can easily be plugged into different work pipelines, with different modeling software. Other alternatives such as Tensorflow-Datasets([@tensorflow_datasets](#)) or HuggingFace Datasets ([@lhoest2021datasets](#)), initially developed for other fields such as Natural Language Processing (NLP) datasets, started including computer vision as well as audio datasets. However, they do not standardize classes, so audio events can come in different formats in each dataset, requiring further work from the user each time. Having a community-centric, open-source, audio-specialized library allows us greater flexibility to incorporate more audio-specific API functionalities and align our priorities with those of the audio community.
@@ -66,7 +66,7 @@ There are other libraries that handle datasets, like `Tensorflow` ([@tensorflow]
 - **Standardize usage of sound datasets:** Standardizes common attributes of sound datasets such as audio or tags to simplify audio research pipelines, while preserving each dataset’s idiosyncrasies (e.g. if a dataset has ‘non-standard’ attributes, we include them as well).
 
 
[…]
