### 2024-02-07 9e4040c
Trim Annotation Types
