+++ paper.bib
@@ -44,13 +44,6 @@ url={https://doi.org/10.48550/arXiv.1605.08695}
   howpublished = {\url{https://www.tensorflow.org/datasets}},
 }
 
-@misc{chollet2015keras,
-  title={Keras},
-  author={Chollet, Fran\c{c}ois and others},
-  year={2015},
-  howpublished={\url{https://keras.io}},
-}
-
 @inproceedings{zinemanas2020dcase,
   title={{DCASE-models:} a phyton library for computational environmental sound analysis using deep-learning models},
   author={Zinemanas, Pablo and Hounie, Ignacio and Cancela, Pablo and Font Corbera, Frederic and Rocamora, Mart{\'\i}n and Serra, Xavier},
