### 2024-02-07 #6331 @magdalenafuentes
Hey @danielskatz, thanks for moving this along. We shorten the paper significantly and included the DOIs, thanks for pointing those out! We're not sure how to check the word count (the tools we're using don't seem to match the word count here), but we did our best to get close to the `example paper`. Please let us know if further changes are needed and thanks again!
