### 2024-02-08 #6331 @xuanxu
I see `Out of memory (Memory quota vastly exceeded)` errors in the server logs, probably that's caused by the size of the repo. I'll try to add some measures.
