### 2024-02-08 #6331 @danielskatz
@magdalenafuentes - @xuanxu reports that our tooling, which isn't currently working consistently due to the size of the repository, produces `Wordcount for paper.md is 1603`.  Can you try to remove a bit more?
