### 2024-02-08 0aa1aca
Update image and reference year
