+++ paper.md
@@ -52,7 +52,7 @@ bibliography: paper.bib
 
 As research pipelines become increasingly complex, it is key that their different components are reproducible. In recent years, the research community has made considerable efforts towards standardization and reproducibility, with modeling and evaluation libraries ([@tensorflow, @pedregosa2011scikit, @mesaros2016metrics](#)), open sourcing models ([@zinemanas2020dcase, @speechbrain](#)) and data dissemination using resources such as [Zenodo](https://zenodo.org). However, discrepancies in the local version of the data and different practices in loading and parsing datasets can lead to considerable differences in performance results, which is misleading when comparing methods ([@bittner2019mirdata](#)). Besides, it is extremely inefficient to develop pipelines from scratch for loading and parsing a dataset for each researcher or team each time, and this increases the chances of bugs and differences that hinder reproducibility.
 
-There are other libraries that handle datasets like `Tensorflow` ([@tensorflow](#)) or Tensorflow-Datasets([@tensorflow_datasets](#)), `DCASE-models` ([@zinemanas2020dcase](#)) or HuggingFace Datasets ([@lhoest2021datasets](#)). But none of them serve as a stand-alone library that can easily be plugged into different work pipelines, with different modeling software.  Having a community-centric, open-source, audio-specialized library allows us greater flexibility to incorporate more audio-specific API functionalities and align our priorities with those of the audio community.
+There are other libraries that handle datasets like `Tensorflow` ([@tensorflow](#)) or Tensorflow-Datasets ([@tensorflow_datasets](#)), `DCASE-models` ([@zinemanas2020dcase](#)) or HuggingFace Datasets ([@lhoest2021datasets](#)). But none of them serve as a stand-alone library that can easily be plugged into different work pipelines, with different modeling software.  Having a community-centric, open-source, audio-specialized library allows us greater flexibility to incorporate more audio-specific API functionalities and align our priorities with those of the audio community.
 
 `Soundata` was created following these design principles:
 
