### 2024-02-08 #6331 @danielskatz
> Hi @danielskatz , we've reduced it to 1121 words (if we don't count the header with authors information and the references the word count goes down to 979). Let us know if that's enough.

Yes, thanks - this looks fine
