### 2024-02-08 #6331 @guillemcortes
Regenerating the pdf since we've modified the figures to avoid unwanted page breaks
