### 2024-02-08 #6331 @faroit
@danielskatz i would love to edit this submission but I would like to verify possible COIs according to [our policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). I know Magdalena from conferences but I can't remember any research or publication we have done together. @magdalenafuentes can you please confirm and check on your side as well, please?
