### 2024-02-09 #6331 @magdalenafuentes
Hey @faroit 👋  I don't think we have directly contributed in any project or publication together, so I don't think we have a COI
