### 2024-02-21 #6331 @faroit
👋  Hi @magdalenafuentes, thanks again your submitting to JOSS! So I'm happy to serve as an editor for this. We'll use this issue to identify reviewers and resolve technical issues.

As sounddata has a wide reach throughout a diverse number of mir tasks, it would be great to have one reviewer with a domain knowledge in one of the tasks but also someone who have experience with data related tools used in the MIR community... I have a few potential reviewers in mind

- __@ fzalkow__
- __@ hadware__
- __@ pzelasko__
- __@ turian__
- __@ nateanl__
- __@ hagenw__

Because open-source audio is a small community, please let me know about potential COIs wrt to the [JOSS policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)

@magdalenafuentes, also let me know about other potential reviewers, so any suggestions you have are also welcome, please only mention them with a space between the @ and their handle (so you would refer to me as "@ faroit").
