### 2024-03-01 #6331 @faroit
👋 @turian @hagenw @nateanl @polarch @danstowell @toni-heittola @keisukeImoto @inesnolas would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>

The software under review is [soundata](https://github.com/soundata/soundata).
