### 2024-03-01 #6331 @hagenw
I would indeed be interested in reviewing this, as I work since a few years on a Python library to manage and version audio data for machine learning as well (https://github.com/audeering/audb).
The only restriction is that I will only have time for the review after 20th March.
