### 2024-03-03 #6331 @faroit
@hagenw I'm sure having you on board makes any software stronger :-) So thanks for your offer and helping out JOSS.
