### 2024-03-27 soundata/soundata#156 @hagenw
How to handle different verisons of a dataset
The [Writing a New Dataset Loader](https://soundata.readthedocs.io/en/latest/source/contributing.html#writing-a-new-dataset-loader) guide does not discuss how to handle datasets that have different versions. At the moment it assumes that each dataset has only a single canonical version, which might not always be the case, e.g. [MSP-Podcast](https://ecs.utdallas.edu/research/researchlabs/msp-lab/MSP-Podcast.html) has several cannonical versions.

Looking at index scripts, I also see that they assign version numbers:

https://github.com/soundata/soundata/blob/62fdaa6e792601c0183766b6bf50574a8d0131db/scripts/make_dcase23_task2_index.py#L14-L18

Is this version the cannonical version?

In addition, I found https://github.com/soundata/soundata/issues/60 which proposes to support multiple versions of a dataset.

---

For now I would propose to expand the contributing guide stating if the support of different versions per dataset is a planned feature. And if not, it should discuss how to handle datasets with different cannonical versions, e.g. incorporating the version into the name like `msp-podcast-1.11`.
