### 2024-03-27 soundata/soundata#157 @hagenw
Use API dataset names in overview table
It is a very nice idea to have an overview table of all supported datasets as provided at https://soundata.readthedocs.io/en/latest/source/quick_reference.html

What might be even more convinient would be to use the same dataset names in the table and in the library.
E.g. when looking at

![image](https://github.com/soundata/soundata/assets/173624/8c677f84-7f83-4594-b7c3-79b56efe290b)

you might try:

```python
>>> import soundata
>>> dataset = soundata.initialize("TAU SSE 2019")
...
ValueError: Invalid dataset TAU SSE 2019
```

instead you need to do:

```python
dataset = soundata.initialize("tau2019sse")
```

It would be great, if a user could just copy and paste the name from the table.
