### 2024-03-27 soundata/soundata#157 @hagenw
It's even more complicated for

![image](https://github.com/soundata/soundata/assets/173624/fee223b8-6dbb-47c1-9d2d-9eb6139a53d2)

where the dataset name is `marco`.
