### 2024-04-15 #6331 @faroit
👋 @fschmid56 @kkoutini @turian @nateanl @polarch @danstowell @toni-heittola @keisukeImoto @inesnolas @gertdekkers @dangpzanco @f0k @bytesnake @liuxubo717 @NielsRogge @hadware we are still looking for a JOSS review of [soundata](https://github.com/soundata/soundata) here on github.

If you can't help out yourself, I would be very helpful if potential reviewers could be proposed with expertise in audio datasets and python (possibly in the DCASE community).

The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
