### 2024-04-15 #6331 @hadware
Sure, I'd happily review this. This looks like a great tool :)
