### 2024-04-16 #6331 @hadware
I've been once on a work trip/holiday in barcelona, that's as far as any CoI my might go (so, yeah, no COI :) )
