### 2024-04-16 #6634 @editorialbot
[REVIEW]: Soundata: Reproducible use of audio datasets
**Submitting author:** @magdalenafuentes (<a href="http://orcid.org/0000-0003-4506-6639">Magdalena Fuentes</a>)
**Repository:** https://github.com/soundata/soundata
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.1
