### 2024-04-17 #6634 @hadware
Hey everyone. Amazing work you did on that library, a real pleasure to be able to review it. I might actually also be using it in the near future :)

@faroit I'm not sure about the authoring: while most authors have strong contributions to the codebase, some (like Guillem Cortès or Xavier Serra) don't have that many in comparison to others. Is that an issue for JOSS? Furthermore, Rachel Bittner from Pysox has made a lot of work on that library, should she be cited as well?
