### 2024-04-17 soundata/soundata#153 @hadware
I agree with this. This is also linked to https://github.com/soundata/soundata/issues/154 and should be solved.
