### 2024-04-17 soundata/soundata#158 @hadware
Fix the packaging
Alright, this package is super well-made and well engineered, but the packaging has some issues:

1.  Basic thing, but I think you should (and can easily) switch to a `pyproject.toml` package file. It's now the [packaging standard for python](https://peps.python.org/pep-0621/). Except for your versioner (i think you can use something from [there](https://packaging.python.org/en/latest/guides/single-sourcing-package-version/), everything should be pretty straightforward
2. As mentioned in https://github.com/soundata/soundata/issues/154 , I doubt that all the dependencies, some of which are pretty bloated, are needed right out of the box. Moreover, some of them aren't installed properly unless you apt-get some package ( https://github.com/soundata/soundata/issues/153 )

I'd advise that you put some of then in some `[project.optional-dependencies]` groups, and list out all the optional optional dependency groups that can be installed, e.g.

```
pip install soundata[plots] # here seaborn & co are installed as well
pip install soundata[playback] # here simpleaudio is installed
# etc, etc
```

The goal is, of course, to make the basic install as light as possible. To prevent any import error, import the relevant package _only at runtime_ and warn the user when the package hasn't been installed (with the instructions on how to fix that).
