### 2024-04-17 soundata/soundata#159 @hadware
Regarding the large repo size
I don't know if this is relevant with my reviewing role for this package, but you might want to clean up the repo and migrate .wav files from tests ans JSON from indexes to LFS : https://docs.github.com/en/repositories/working-with-files/managing-large-files/moving-a-file-in-your-repository-to-git-large-file-storage
