### 2024-04-17 soundata/soundata#160 @hadware
Type hinting
I've seen that some of this package's code has a bit of type-hinting.

It'd be really nice if at least the user-facing functions, classes's `__init__` and methods have a bit of type hinting. This makes things a bit clearer for the user (even though you already have extensive docstrings which do carry that information). In my case for instance, my editor is very much type-aware and having type-hints is always extremely helpful to understand what kind of function argument or output i'm dealing with.

I'd only worry about type hinting stuff in the root `soundata/` (not all the classes from `datasets`) folder for instance. For classes that do very little except for holding a buch of data (i'm looking at you, `annotation.py`), maybe just type-hint the class attributes).
