### 2024-04-17 #6634 @magdalenafuentes
Hi @hadware!

Thanks for reviewing this 🙏

As @faroit said I'm currently on leave so I won't be able to follow the reviewing process as much, but maybe I can quickly clarify this: senior authors such as Xavier Serra and Juan Bello helped in the conceptualization of the libraries (e.g. deciding directions/datasets to focus on) as well as funding the team. @rabitt was initially in the authors list as we jointly created mirdata (the music cousin of soundata), but due to issues with her employer she had to remove herself, and she knows and is onboard with us continuing the project without her. @faroit can clarify further if this presents any issues with JOSS policies.
