### 2024-05-03 soundata/soundata#157 @genisplaja
Hello @hagenw, thank you very much for your review. We are working on the raised points, and we will be getting back to you soon. About this one, we agree with your point here. Our preferred solution would be to add an extra column to the dataset table indicating the API dataset name (e.g. `marco`) alongside its official formatted name (e.g. `3D-MARCo`). Therefore, we would have the complete dataset name for more information, while indicating the API dataset id to use to initialize the dataset, as some dataset ids ay skip some important parts of the name. What do you think about that? Thanks!
