### 2024-05-06 soundata/soundata#159 @hagenw
The problem regarding the JSON files will only be solved partially by using LFS, as they also need to be included in the Python package, see https://github.com/soundata/soundata/issues/155 for the proposal to separate the index from the Python package (which would also remove the files from this repository).
