### 2024-05-06 soundata/soundata#161 @hagenw
Add a CONTRIBUTING file to the root of the repository
To make it easier to find the documentation on contributing, I would recommend to add a `CONTRIBUTING.rst` to the root of the repo.
It's content can then simply be included inside `docs/source/contributing.rst`, e.g.

```rst
.. include:: ../../CONTRIBUTING.rst
```
