### 2024-05-06 soundata/soundata#162 @hagenw
Include the command how to build documentation locally
At the moment it is discussed in the contribution documentation, that the documentation is automatically build for each pull request, which is a very nice feature. Nonetheless, I would also recommend to describe how to build the documentation locally.

For example, I tried to build it with running:
```bash
$ python -m sphinx docs/ build/html -b html
```
The build succeeded, but I got 134 warnings. So I'm not sure if I should have done it differently?
