### 2024-05-06 soundata/soundata#159 @guillemcortes
Hi @hadware @hagenw, thank you for your valuable feedback. We've been investigating and it looks like the best option will be to use [Zenodo Communities](https://zenodo.org/communities) to store both indices and the resources files for the tests (.wav files, mostly) so they can be downloaded like it is done in [mirdata](https://github.com/mir-dataset-loaders/mirdata/blob/master/mirdata/datasets/acousticbrainz_genre.py#L68). This would add an extra step to the contributors since they will have to upload the files there, but we believe that is a sustainable solution. We are testing now how it would work. Do you have experience with Zenodo Communities? Can you think of any issues we might be overlooking? Thanks again!
