### 2024-05-06 soundata/soundata#163 @hagenw
Include Rachel Bittner to the list of authors in the JOSS paper
Looking at [the contributions of Rachel Bittner](https://github.com/soundata/soundata/commits?author=rabitt), they include work on the basic package, and do not only add datasets, as for example [Iran R. Roman](https://github.com/soundata/soundata/commits?author=iranroman) did.
