### 2024-05-06 #6634 @hagenw
I would propose to include Rachel Bittner to the list of authors, see https://github.com/soundata/soundata/issues/163
