### 2024-05-06 #6634 @faroit
> I would propose to include Rachel Bittner to the list of authors, see https://github.com/soundata/soundata/issues/163

@hagenw thanks for bringing this up again but I think we have to accepts Magdalenas comment here https://github.com/openjournals/joss-reviews/issues/6634#issuecomment-2062315446
