### 2024-05-06 soundata/soundata#159 @hagenw
I think regarding the files needed for testing, I don't see an issue with using LFS and store the files inside this repo.
For the indices, it would indeed be better if they are hosted on a server from which they can be downloaded on the fly. Of course, there are not so many great and free options available for this. I have used Zenodo for this, but it could also lead to some problems, compare https://github.com/sigsep/open-unmix-pytorch/issues/138. I don't have experience with Zenodo Communities, but yes, it sounds like a possible solution.
