### 2024-05-06 soundata/soundata#159 @hagenw
In the documentation, there seems to be already an example for hosting an index on Zenodo: https://soundata.readthedocs.io/en/latest/source/contributing.html#working-with-remote-indexes
