### 2024-05-06 soundata/soundata#164 @hagenw
Extend documentation on how to contribute a new dataset
There is already extensive documentation on how to contribute a new dataset: https://soundata.readthedocs.io/en/latest/source/contributing.html#writing-a-new-dataset-loader

This together with the source-code of the existing datasets, was enough for me to create a new loader for an example dataset.

But in the JOSS paper it is written:

> Easy to contribute to: Users do not need to go through all the source code to contribute.
> Soundata provides extensive documentation explaining how to contribute a new loader.

This, I cannot confirm as some details are missing in the documentation.
If the statement above should be met, I would propose to extend the documentation by:

* Discuss if I'm allowed to use packages inside the index generation script, that are not part of the dependencies of `sounddata`
* Provide better explanation for handling of datasets, that are not stored in the clip based approach
* Discuss why I should use `download_utils.RemoteFileMetadata` in the data loader and not something like `download_utils.download_zip_file()`
* Mention that the current workflow is not well suited if the original data provides a single ZIP file for every audio file
* Discuss how folds/splits should be handled, e.g. adding a `split` property
* Discuss that the documentation of the dataset should be placed in the docstring of the data loader. State, how the docstring is supposed to look like, e.g. which sub-section should be included
* State where and how the license, authors, bibtex reference of the dataset should be mentioned
