### 2024-05-06 #6634 @hagenw
I have finished the review. It's a very nice contribution to the audio community.

Here is a list of the issues, I would like to be tackled before accepting all points in the bullet list:

* [Extend documentation on how to contribute a new dataset #164](https://github.com/soundata/soundata/issues/164)
* [Include the command how to build documentation locally #162](https://github.com/soundata/soundata/issues/162)
* [Add a CONTRIBUTING file to the root of the repository #161](https://github.com/soundata/soundata/issues/161)
* [Expand installation instructions #153](https://github.com/soundata/soundata/issues/153)

In addition, I would highly recommend to work on the following issues:

* [Fix the packaging #158](https://github.com/soundata/soundata/issues/158)
* [Regarding the large repo size #159](https://github.com/soundata/soundata/issues/159)
* [Use API dataset names in overview table #157](https://github.com/soundata/soundata/issues/157)
* [How to handle different verisons of a dataset #156](https://github.com/soundata/soundata/issues/156)
