+++ docs/source/contributing.rst
@@ -124,6 +124,9 @@ Here's an example of an index to use as a guide:
 
 More examples of scripts used to create dataset indexes can be found in the `scripts <https://github.com/soundata/soundata/tree/master/scripts>`_ folder.
 
+    .. note::
+        Users should be able to create the dataset indexes without the need for additional dependencies that are not included in soundata by default. Should you need an additional dependency for a specific reason, please open an issue to discuss with the maintainer team the need for it.
+
 Example index with clips
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
@@ -131,6 +134,15 @@ Most sound datasets are organized as a collection of clips and annotations. In s
 top-level key. Under this ``clips`` top-level key, you should store a dictionary where the keys are the unique clip ids of the dataset, and
 the values are dictionaries of files associated with a clip id, along with their checksums. These files can be for instance audio files
 or annotations related to the clip id. File paths are relative to the top level directory of a dataset.
+
+    .. note::
+        If your sound dataset does not fit into a structure around the clip class, please open an issue in the GitHub repository to discuss how to proceed. These are corner cases that we address especially to maintain the consistency of the library.
+
+Currently, soundata does not include built-in functions to automatically create train, testing, and validation splits if these are not originally defined in the dataset. 
+Users can do that using  external functions such as ``sklearn.model_selection.train_test_split``.
+If a dataset has predefined splits, you can include the split name as an attribute of the ``Clip`` class. You should not create separate indexes for the different splits, or indicate the split in the index.
+See an example of how an index should look like:
+
 
 .. admonition:: Index Examples - Clips
     :class: dropdown
@@ -157,45 +169,46 @@ or annotations related to the clip id. File paths are relative to the top level 
     .. code-block:: javascript
 
 
-        {   "version": "1.0",
-            "clips":
-                "clip1": {
-                    "audio": [
-                        "audio/clip1.wav",  // the relative path for clip1's audio file
-                        "912ec803b2ce49e4a541068d495ab570"  // clip1.wav's md5 checksum
-                    ],
-                    "annotation": [
-                        "annotations/clip1.csv",  // the relative path for clip1's annotation
-                        "2cf33591c3b28b382668952e236cccd5"  // clip1.csv's md5 checksum
+        {
+            "version": "1.0",
+                "clips":
+                    "clip1": {
+                        "audio": [
+                            "audio/clip1.wav",  // the relative path for clip1's audio file
+                            "912ec803b2ce49e4a541068d495ab570"  // clip1.wav's md5 checksum
+                        ],
+                        "annotation": [
+                            "annotations/clip1.csv",  // the relative path for clip1's annotation
+                            "2cf33591c3b28b382668952e236cccd5"  // clip1.csv's md5 checksum
+                        ]
+                    },
+                    "clip2": {
+                        "audio": [
+                            "audio/clip2.wav",
+                            "65d671ec9787b32cfb7e33188be32ff7"
+                        ],
+                        "annotation": [
+                            "annotations/Clip2.csv",
+                            "e1964798cfe86e914af895f8d0291812"
+                        ]
+                    },
+                    "clip3": {
+                        "audio": [
+                            "audio/clip3.wav",
+                            "60edeb51dc4041c47c031c4bfb456b76"
+                        ],
+                        "annotation": [
+                            "annotations/clip3.csv",
+                            "06cb006cc7b61de6be6361ff904654b3"
+                        ]
+                    },
+                }
+            "metadata": {
+                    "metadata_file": [
+                        "metadata/metadata_file.csv",
+                        "7a41b280c7b74e2ddac5184708f9525b"
                     ]
-                },
-                "clip2": {
[…]
