### 2024-05-06 soundata/soundata#161 @genisplaja
Dear @hagenw, thank you for your suggestion. We've tested this one, but GitHub is not [rendering the notes properly](https://github.com/github/markup/issues/1682) and it doesn't properly import files as sphinx does, therefore, to enchance readability, we should be creating a standalone contribution file for the root folder, and another standalone file for the docs, and we believe that this may be a bit overkill.

In fact, in the `README.md` file we are linking directly the contribution guidelines. Maybe we would propose to move this up and highlight it, for users to find it quickly. Does that sound good? Thanks!
