### 2024-05-07 soundata/soundata#161 @hagenw
Yes, good point.

An alternative solution would be to move parts from the "Contributing" section to their own top-level section, but I think the current structure makes sense, and I would propose to stay with it.

So maybe add a short sentence to the top of the README as well, e.g. inside

![image](https://github.com/soundata/soundata/assets/173624/7d589063-bd55-426f-88cf-e843966afd6e)

I also like that you have only two lines there, so maybe try to fit it in the first line? E.g.

> Python library for downloading & working with sound datasets. Check the [API documentation]() and [contribute]() datasets.
