### 2024-05-07 soundata/soundata#159 @nkundiushuti
I think we excluded LFS because of the size limitation (2GB per repo?). in the other repo (mirdata) we had issues with some very large indexes. to sum up, it's not sustainable to store them on LFS and zenodo is a better option.
