### 2024-05-07 fd091ee
supress warnings and errors in docs
