+++ docs/source/contributing.rst
@@ -1,3 +1,790 @@
 .. _contributing:
 
-.. include:: ../../CONTRIBUTING.rst
+########################
+Contributing to soundata
+########################
+
+We encourage contributions to soundata, especially new dataset loaders. To contribute a new loader, follow the
+steps indicated below and create a Pull Request (PR) to the github repository. For any doubt or comment about
+your contribution, you can always submit an issue or open a discussion in the repository.
+
+- `Issue Tracker <https://github.com/soundata/soundata/issues>`__
+- `Source Code <https://github.com/soundata/soundata>`__
+
+Quick link to contributing templates
+####################################
+
+If you're familiar with Soundata's API already, you can find the template files for contributing `here <https://github.com/soundata/soundata/tree/master/docs/source/contributing_examples>`__,
+and the loader checklist for submiting your PR `here <https://github.com/soundata/soundata/blob/master/.github/PULL_REQUEST_TEMPLATE/new_loader.md>`__.
+
+
+
+
+Installing soundata for development purposes
+############################################
+
+To install ``soundata`` for development purposes:
+
+    - First run:
+
+    .. code-block:: console
+
+        git clone https://github.com/soundata/soundata.git
+
+    - Then, after opening source data library you have to install the dependencies for updating the documentation
+      and running tests:
+
+    .. code-block:: console
+
+        pip install .
+        pip install ."[tests]"
+        pip install ."[docs]"
+
+
+We recommend using `miniconda <https://docs.conda.io/en/latest/miniconda.html>`__ or
+`pyenv <https://github.com/pyenv/pyenv#installation>`__ to manage your Python versions
+and install all ``soundata`` requirements. You will want to install the latest supported Python versions (see README.md).
+Once ``conda`` or ``pyenv`` and the Python versions are configured, install ``pytest``. Make sure you've installed all the 
+necessary pytest plugins needed (e.g. `pytest-cov`) to automatically test your code successfully.
+
+Before running the tests, make sure to have formatted ``soundata/`` and ``tests/`` with ``black``.
+
+.. code-block:: bash
+
+    black soundata/ tests/
+
+
+Also, make sure that they pass flake8 and mypy tests specified in lint-python.yml github action workflow.
+
+.. code-block:: bash
+
+    flake8 soundata --count --select=E9,F63,F7,F82 --show-source --statistics
+    python -m mypy soundata --ignore-missing-imports --allow-subclassing-any
+
+
+Finally, run:
+
+.. code-block:: bash
+
+    pytest -vv --cov-report term-missing --cov-report=xml --cov=soundata tests/ --local
+
+
+All tests should pass!
+
+.. note::
+        Soundata assumes that your system has the zip library installed for unzipping files. 
+
+
+Writing a new dataset loader
+############################
+
+
+The steps to add a new dataset loader to ``soundata`` are:
+
+1. `Create an index <create_index_>`_
+2. `Create a module <create_module_>`_
+3. `Add tests <add_tests_>`_
+4. `Submit your loader <submit_loader_>`_
+
+**Before starting**, if your dataset **is not fully downloadable** you should:
+
+
+1. Contact the soundata team by opening an issue or PR so we can discuss how to proceed with the closed dataset.
+2. Show that the version used to create the checksum is the "canonical" one, either by getting the version from the 
+   dataset creator, or by verifying equivalence with several other copies of the dataset.
+
+To reduce friction, we will make commits on top of contributors PRs by default unless
+the ``please-do-not-edit`` flag is used.
+
+.. _create_index:
+
+1. Create an index
+------------------
+
+Soundata's structure relies on ``indexes``. Indexes are dictionaries that contain information about the structure of the
+dataset which is necessary for the loading and validating functionalities of Soundata. In particular, indexes contain
[…]
