### 2024-05-07 6af6fba
supress warnings and errors in docs 2
