### 2024-05-07 5ade305
remove unused doc source file
