### 2024-05-07 soundata/soundata#162 @genisplaja
No, you are doing it right. It could actually be done running `make clean html` also. The warnings are produced by formatting issues in the docstrings of the data loaders. I've been fixing them all, and wrote down in the documentation some indications to prevent these warnings. Thanks for bringing this up!
