### 2024-05-07 027574c
add comments on doc error checking
