### 2024-05-07 cfa24b9
make contributing link more visible in README
