+++ soundata/datasets/tau2021sse_nigens.py
@@ -232,7 +232,7 @@ def load_audio(fhandle: BinaryIO, sr=24000) -> Tuple[np.ndarray, float]:
 @io.coerce_to_string_io
 def load_spatialevents(fhandle: TextIO, dt=0.1) -> annotations.SpatialEvents:
     """Load an TAU NIGENS SSE 2021 annotation file
-    
+
     Args:
         fhandle (str or file-like): File-like object or path to
             the sound events annotation file
