### 2024-05-07 72d9d2c
WIP setup to pyproject.toml
