### 2024-05-08 e1a78d4
latex formatted citations and indentation problems
