+++ soundata/datasets/dcase23_task6b.py
@@ -5,14 +5,16 @@ DCASE 2023 Task-6B Dataset Loader
     :class: dropdown
 
     **DCASE 2023 Task-6B**
-    | Clotho (c) by K. Drossos, S. Lipping, and T. Virtanen.
-    | Clotho is licensed under the terms set by Tampere University and Creative Commons licenses for the audio files as per their origin from the Freesound platform.
-    | You should have received a copy of the license along with this work.
-    | https://github.com/audio-captioning/clotho-dataset. Paper: "Clotho: an Audio Captioning Dataset," ICASSP 2020
+
+        | Clotho (c) by K. Drossos, S. Lipping, and T. Virtanen.
+        | Clotho is licensed under the terms set by Tampere University and Creative Commons licenses for the audio files as per their origin from the Freesound platform.
+        | You should have received a copy of the license along with this work.
+        | https://github.com/audio-captioning/clotho-dataset. Paper: "Clotho: an Audio Captioning Dataset," ICASSP 2020
 
     *Created By:*
-    | K. Drossos, S. Lipping, and T. Virtanen.
-    |  Tampere University, Finland
+
+        | K. Drossos, S. Lipping, and T. Virtanen.
+        |  Tampere University, Finland
 
     *Version 2.1.0*
         Fixes for corrupted files and illegal characters.
@@ -47,7 +49,10 @@ DCASE 2023 Task-6B Dataset Loader
         Full details in the LICENSE file included with the dataset.
 
     *Acknowledgment in Academic Research*
-        When using Clotho for academic research, please cite:
+    When using Clotho for academic research, please cite:
+        
+    .. code-block:: latex
+
         K. Drossos, S. Lipping, and T. Virtanen, "Clotho: an Audio Captioning Dataset," ICASSP 2020.
 
     *Feedback and Contributions*
