### 2024-05-08 15b5bd3
Move requests and pydub to optional dependencies
