+++ pyproject.toml
@@ -31,17 +31,16 @@ dependencies = [
     "librosa>=0.10.0",
     "numpy>=1.21.6",
     "pandas>=1.3.5",
-    "requests>=2.31.0",
     "tqdm>=4.65.0",
     "jams>=0.3.4",
     "flake8>=5.0.4",
     "mypy>=0.982",
     "py7zr>=0.16.0",
-    "pydub>=0.25.1",
 ]
 
 [project.optional-dependencies]
 tests = [
+    "requests>=2.31.0",
     "pytest>=7.2.0",
     "pytest-cov>=4.1.0",
     "pytest-pep8>=1.0.6",
@@ -65,6 +64,7 @@ plots = [
     "simpleaudio>=1.0.4",
     "seaborn>=0.11.2",
     "ipywidgets>=8.1.1",
+    "pydub>=0.25.1",
 ]
 
 [project.urls]
