### 2024-05-08 525293b
Updating installation documentation
