### 2024-05-10 soundata/soundata#164 @genisplaja
hello @hagenw! Thanks again for your review. In our last PR (#165) we addressed the concerns in issues #161, #162, and the one here. We internally reviewed that everything requested was addressed there, and merged in master. Please feel free to revise that the updates are satisfactory in #165.

Other than that, for future updates following the review comments, please do let us know if we should include you and @hadware as reviewers for the PR, or if you prefer to re-review after we've made all fixes.

Thanks!
