### 2024-05-12 soundata/soundata#157 @guillemcortes
Hi @hagenw, we've merged #167 which addresses this issue. You can check how it looks here: https://soundata.readthedocs.io/en/latest/source/quick_reference.html
