+++ soundata/datasets/freefield1010.py
@@ -72,6 +72,11 @@ REMOTES = {
         url="https://ndownloader.figshare.com/files/10853303",
         checksum="2f4e7e194ccbd3de86e997af8f2a0405",
     ),
+    "index": download_utils.RemoteFileMetadata(
+        filename="freefield1010_index_1.0.json",
+        url="https://zenodo.org/records/11176811/files/freefield1010_index_1.0.json?download=1",
+        checksum="0aed8eebb8fb66da1a8437729dfa0f05",
+    )
 }
 
 LICENSE_INFO = "Creative Commons Attribution Non Commercial 4.0 International"
@@ -183,6 +188,7 @@ class Dataset(core.Dataset):
             bibtex=BIBTEX,
             remotes=REMOTES,
             license_info=LICENSE_INFO,
+            custom_index_path="freefield1010_index_1.0.json"
         )
 
     @core.copy_docs(load_audio)
