+++ soundata/datasets/tau2020sse_nigens.py
@@ -134,6 +134,11 @@ REMOTES = {
         url="http://zenodo.org/record/4064792/files/metadata_eval.zip?download=1",
         checksum="f3584166d9a63b43c1e301b6fb722293",
     ),
+    "index": download_utils.RemoteFileMetadata(
+        filename="tau2020sse_nigens_index_1.2.0.json",
+        url="https://zenodo.org/records/11176861/files/tau2020sse_nigens_index_1.2.0.json?download=1",
+        checksum="3ec732e50263d352070a32bc4af73eb1",
+    )
 }
 
 LICENSE_INFO = """
@@ -399,6 +404,7 @@ class Dataset(core.Dataset):
             bibtex=BIBTEX,
             remotes=REMOTES,
             license_info=LICENSE_INFO,
+            custom_index_path="tau2020sse_nigens_index_1.2.0.json",
         )
 
     @core.copy_docs(load_audio)
