+++ soundata/datasets/tut2017se.py
@@ -236,6 +236,11 @@ REMOTES = {
         ),
         checksum="a951598abaea87296ca409e30fb0b379",
     ),
+    "index": download_utils.RemoteFileMetadata(
+        filename="tut2017se_index_2.0.json",
+        url="https://zenodo.org/records/11176916/files/tut2017se_index_2.0.json?download=1",
+        checksum="26fea2fd4082f48d7e8ef8c85df88ad1",
+    )
 }
 
 LICENSE_INFO = "TUT License <https://github.com/TUT-ARG/DCASE2017-baseline-system/blob/master/EULA.pdf>"
@@ -382,6 +387,7 @@ class Dataset(core.Dataset):
             bibtex=BIBTEX,
             remotes=REMOTES,
             license_info=LICENSE_INFO,
+            custom_index_path="tut2017se_index_2.0.json",
         )
 
     @core.copy_docs(load_audio)
