### 2024-05-13 soundata/soundata#161 @hagenw
This was solved by #165
