### 2024-05-13 soundata/soundata#162 @hagenw
I realized, I overlooked the documentation on how to build the docs before. Maybe it is not optimal, that the documentation is discussed at two different places in contributing, namely "Writing a new dataset loader > 4. Submit your loader > Docs" and under "Documentation".

I can confirm that with #165 I don't get any warnings when building the docs locally.
