### 2024-05-13 soundata/soundata#164 @hagenw
Thanks, you addressed indeed everything in #165.

Feel free to include me as a reviewer in further PRs, if you like to get some feedback.
