### 2024-05-14 fb60496
Unifying documentation tutorial in contributing docs (#170)
* unifying documentation info

* addressing @guillemcortes's review

* missing space and consistent lists
