### 2024-05-14 da55559
Modify example.py and test_example.py
