+++ docs/source/contributing_examples/example.py
@@ -34,7 +34,20 @@ BIBTEX = """
   year = "1986"
 }
 """
-
+# -- INDEXES specifies different versions of a dataset
+# -- "default" and "test" specify which key should be used
+# -- by default, and when running tests.
+# -- Some datasets have a "sample" version, which is a mini-version
+# -- that makes it easier to try out a large dataset without needing
+# -- to download the whole thing.
+# -- If there is no sample version, simply set "test": "1.0".
+# -- If the default data is remote, there must be a local sample for tests!
+INDEXES = {
+    "default": "1.0",
+    "test": "sample",
+    "1.0": core.Index(filename="example_index_1.0.json"),
+    "sample": core.Index(filename="example_index_sample.json")
+}
 # -- REMOTES is a dictionary containing all files that need to be downloaded.
 # -- The keys should be descriptive (e.g. 'annotations', 'audio').
 # -- When having data that can be partially downloaded, remember to set up
@@ -169,12 +182,14 @@ def load_annotation(fhandle):
 class Dataset(core.Dataset):
     """The Example dataset"""
 
-    def __init__(self, data_home=None):
+    def __init__(self, data_home=None, version="default"):
         super().__init__(
             data_home,
+            version,
             name='dataset_name',
             clip_class=Clip,
             bibtex=BIBTEX,
+            indexes=INDEXES,
             remotes=REMOTES,
             download_info=DOWNLOAD_INFO,
             license_info=LICENSE_INFO,
