+++ tests/test_download_utils.py
@@ -4,7 +4,7 @@ import shutil
 import zipfile
 import re
 
-from soundata import download_utils
+from soundata import download_utils, core
 from soundata.datasets import esc50
 
 import pytest
@@ -72,48 +72,49 @@ def test_downloader(mocker, mock_path):
     file_remote = download_utils.RemoteFileMetadata(
         filename="remote.txt", url="a", checksum=("1234")
     )
+    index = core.Index("asdf.json")
 
     # Zip only
-    download_utils.downloader("a", remotes={"b": zip_remote})
+    download_utils.downloader("a", index=index, remotes={"b": zip_remote})
     mock_zip.assert_called_once_with(zip_remote, "a", False, False)
     mocker.resetall()
 
     # tar only
-    download_utils.downloader("a", remotes={"b": tar_remote})
+    download_utils.downloader("a", index=index, remotes={"b": tar_remote})
     mock_tar.assert_called_once_with(tar_remote, "a", False, False)
     mocker.resetall()
 
     # 7z only
-    download_utils.downloader("a", remotes={"b": _7z_remote})
+    download_utils.downloader("a", index=index, remotes={"b": _7z_remote})
     mock_7z.assert_called_once_with(_7z_remote, "a", False, False)
     mocker.resetall()
 
     # file only
-    download_utils.downloader("a", remotes={"b": file_remote})
+    download_utils.downloader("a", index=index, remotes={"b": file_remote})
     mock_download_from_remote.assert_called_once_with(file_remote, "a", False)
     mocker.resetall()
 
     # zip and tar
-    download_utils.downloader("a", remotes={"b": zip_remote, "c": tar_remote})
+    download_utils.downloader("a", index=index, remotes={"b": zip_remote, "c": tar_remote})
     mock_zip.assert_called_once_with(zip_remote, "a", False, False)
     mock_tar.assert_called_once_with(tar_remote, "a", False, False)
     mocker.resetall()
 
     # zip and file
-    download_utils.downloader("a", remotes={"b": zip_remote, "c": file_remote})
+    download_utils.downloader("a", index=index, remotes={"b": zip_remote, "c": file_remote})
     mock_zip.assert_called_once_with(zip_remote, "a", False, False)
     mock_download_from_remote.assert_called_once_with(file_remote, "a", False)
     mocker.resetall()
 
     # tar and file
-    download_utils.downloader("a", remotes={"b": tar_remote, "c": file_remote})
+    download_utils.downloader("a", index=index, remotes={"b": tar_remote, "c": file_remote})
     mock_tar.assert_called_once_with(tar_remote, "a", False, False)
     mock_download_from_remote.assert_called_once_with(file_remote, "a", False)
     mocker.resetall()
 
     # zip and tar and file
     download_utils.downloader(
-        "a", remotes={"b": zip_remote, "c": tar_remote, "d": file_remote}
+        "a", index=index, remotes={"b": zip_remote, "c": tar_remote, "d": file_remote}
     )
     mock_zip.assert_called_once_with(zip_remote, "a", False, False)
     mock_download_from_remote.assert_called_once_with(file_remote, "a", False)
@@ -121,13 +122,14 @@ def test_downloader(mocker, mock_path):
     mocker.resetall()
 
     # Zip multipart
-    download_utils.downloader("a", remotes={"b": multipart_zip_remote})
+    download_utils.downloader("a", index=index, remotes={"b": multipart_zip_remote})
     mock_multipart_zip.assert_called_once_with(multipart_zip_remote, "a", False, False)
     mocker.resetall()
 
     # test partial download
[…]
