+++ tests/test_full_dataset.py
@@ -12,13 +12,13 @@ import soundata
 
 
 @pytest.fixture()
-def dataset(test_dataset):
+def dataset(test_dataset, dataset_version):
     if test_dataset == "":
         return None
     elif test_dataset not in soundata.DATASETS:
         raise ValueError("{} is not a dataset in soundata".format(test_dataset))
     data_home = os.path.join("tests/resources/sound_datasets_full", test_dataset)
-    return soundata.initialize(test_dataset, data_home)
+    return soundata.initialize(test_dataset, data_home, version=dataset_version)
 
 
 # This is magically skipped by the the remote fixture `skip_remote` in conftest.py

[…]
