### 2024-05-14 e4924f1
documentation for display_plot
