+++ soundata/datasets/urbansed.py
@@ -157,7 +157,7 @@ REMOTES = {
         filename="urbansed_index_2.0.json",
         url="https://zenodo.org/records/11176925/files/urbansed_index_2.0.json?download=1",
         checksum="95e7a152729ebdb386356e7d21b342b7",
-    )
+    ),
 }
 
 LICENSE_INFO = "Creative Commons Attribution 4.0 International"
