### 2024-05-14 5f445b0
documentation for display_plot
