### 2024-05-14 4e22519
documentation for display_plot
