### 2024-05-14 f0aac83
dcase_birdVox20k example
