### 2024-05-14 7a25db9
WIP fixing core.py and test_core.py
