+++ tests/datasets/test_eigenscape.py
@@ -12,7 +12,7 @@ TEST_DATA_HOME = os.path.normpath("tests/resources/sound_datasets/eigenscape")
 
 def test_clip():
     default_clipid = "Beach.1"
-    dataset = eigenscape.Dataset(TEST_DATA_HOME)
+    dataset = eigenscape.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
 
     expected_attributes = {
@@ -36,7 +36,7 @@ def test_clip():
 
 def test_load_audio():
     default_clipid = "Beach.1"
-    dataset = eigenscape.Dataset(TEST_DATA_HOME)
+    dataset = eigenscape.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     audio_path = clip.audio_path
     audio, sr = eigenscape.load_audio(audio_path)
@@ -50,7 +50,7 @@ def test_load_audio():
 def test_load_tags():
     # dataset
     default_clipid = "Beach.1"
-    dataset = eigenscape.Dataset(TEST_DATA_HOME)
+    dataset = eigenscape.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     assert len(clip.tags.labels) == 1
     assert clip.tags.labels[0] == "Beach"
@@ -60,7 +60,7 @@ def test_load_tags():
 def test_load_metadata():
     # dataset
     default_clipid = "Beach.1"
-    dataset = eigenscape.Dataset(TEST_DATA_HOME)
+    dataset = eigenscape.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     assert clip.location == "Bridlington Beach"
     assert clip.time == "10:42"
@@ -70,7 +70,7 @@ def test_load_metadata():
 
 def test_to_jams():
     default_clipid = "Beach.1"
-    dataset = eigenscape.Dataset(TEST_DATA_HOME)
+    dataset = eigenscape.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     jam = clip.to_jams()
 
