### 2024-05-16 4ee89c8
flake8 and mypy as optional dependencies
