+++ docs/requirements.txt
@@ -1,15 +1,0 @@
-sphinx-togglebutton
-jams
-librosa>=0.7.0
-tqdm
-sphinx==5.0.0
-pandas
-scipy  #==1.11.1
-numpy  #==1.25.2
-matplotlib  #==3.7.2
-flake8
-sphinx_rtd_theme
-py7zr>=0.16.0
-pydub
-ipywidgets
-seaborn
