### 2024-05-16 5a383cd
Remove install plots optional dependencies from development guidelines
