+++ docs/source/contributing.rst
@@ -39,7 +39,6 @@ To install ``soundata`` for development purposes:
         pip install .
         pip install ."[tests]"
         pip install ."[docs]"
-        pip install ."[plots]"
 
 
 We recommend using `miniconda <https://docs.conda.io/en/latest/miniconda.html>`__ or
