### 2024-05-17 2e6ab53
Fix documentation display_plot_utils
