+++ tests/datasets/test_starss2022.py
@@ -12,7 +12,7 @@ TEST_DATA_HOME = os.path.normpath("tests/resources/sound_datasets/starss2022")
 
 def test_clip():
     default_clipid = "foa_dev/dev-train-sony/fold3_room21_mix001"
-    dataset = starss2022.Dataset(TEST_DATA_HOME)
+    dataset = starss2022.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
 
     expected_attributes = {
@@ -39,7 +39,7 @@ def test_clip():
 
 
 def test_load_audio():
-    dataset = starss2022.Dataset(TEST_DATA_HOME)
+    dataset = starss2022.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip("foa_dev/dev-train-sony/fold3_room21_mix001")
     audio_path = clip.audio_path
     audio, sr = starss2022.load_audio(audio_path)
@@ -51,7 +51,7 @@ def test_load_audio():
 
 
 def test_load_SpatialEvents():
-    dataset = starss2022.Dataset(TEST_DATA_HOME)
+    dataset = starss2022.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip("foa_dev/dev-train-sony/fold3_room21_mix001")
     annotations_path = clip.csv_path
     starss_annotations = starss2022.load_spatialevents(annotations_path)
@@ -127,7 +127,7 @@ def test_load_SpatialEvents():
 
 def test_to_jams():
     default_clipid = "foa_dev/dev-train-sony/fold3_room21_mix001"
-    dataset = starss2022.Dataset(TEST_DATA_HOME)
+    dataset = starss2022.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     jam = clip.to_jams()
 
