+++ soundata/datasets/tut2017se.py
@@ -179,6 +179,18 @@ BIBTEX = """
              System},
     Year = {2017}}
 """
+
+INDEXES = {
+    "default": "2.0",
+    "test": "sample",
+    "2.0": core.Index(
+        filename="tut2017se_index_2.0.json",
+        url="https://zenodo.org/records/11176916/files/tut2017se_index_2.0.json?download=1",
+        checksum="26fea2fd4082f48d7e8ef8c85df88ad1",
+    ),
+    "sample": core.Index(filename="tut2017se_index_2.0_sample.json")
+}
+
 REMOTES = {
     "development.audio.1": download_utils.RemoteFileMetadata(
         filename="TUT-sound-events-2017-development.audio.1.zip",
@@ -235,12 +247,7 @@ REMOTES = {
             "evaluation.meta.zip?download=1"
         ),
         checksum="a951598abaea87296ca409e30fb0b379",
-    ),
-    "index": download_utils.RemoteFileMetadata(
-        filename="tut2017se_index_2.0.json",
-        url="https://zenodo.org/records/11176916/files/tut2017se_index_2.0.json?download=1",
-        checksum="26fea2fd4082f48d7e8ef8c85df88ad1",
-    ),
+    )
 }
 
 LICENSE_INFO = "TUT License <https://github.com/TUT-ARG/DCASE2017-baseline-system/blob/master/EULA.pdf>"
@@ -379,15 +386,16 @@ def load_events(fhandle: TextIO) -> annotations.Events:
 class Dataset(core.Dataset):
     """The TUT Sound events 2017 dataset"""
 
-    def __init__(self, data_home=None):
+    def __init__(self, data_home=None, version="default"):
         super().__init__(
             data_home,
+            version,
             name="tut2017se",
             clip_class=Clip,
             bibtex=BIBTEX,
+            indexes=INDEXES,
             remotes=REMOTES,
             license_info=LICENSE_INFO,
-            custom_index_path="tut2017se_index_2.0.json",
         )
 
     @core.copy_docs(load_audio)
