+++ tests/datasets/test_urbansound8k.py
@@ -13,7 +13,7 @@ TEST_DATA_HOME = os.path.normpath("tests/resources/sound_datasets/urbansound8k")
 
 def test_clip():
     default_clipid = "135776-2-0-49"
-    dataset = urbansound8k.Dataset(TEST_DATA_HOME)
+    dataset = urbansound8k.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
 
     expected_attributes = {
@@ -41,7 +41,7 @@ def test_clip():
 
 
 def test_load_audio():
-    dataset = urbansound8k.Dataset(TEST_DATA_HOME)
+    dataset = urbansound8k.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip("135776-2-0-49")
     audio_path = clip.audio_path
     audio, sr = urbansound8k.load_audio(audio_path)
@@ -54,7 +54,7 @@ def test_load_audio():
 def test_to_jams():
     # Note: original file is 4 sec, but for testing we've trimmed it to 1 sec
     default_clipid = "135776-2-0-49"
-    dataset = urbansound8k.Dataset(TEST_DATA_HOME)
+    dataset = urbansound8k.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     jam = clip.to_jams()
 
