+++ tests/datasets/test_marco.py
@@ -11,7 +11,7 @@ TEST_DATA_HOME = os.path.normpath("tests/resources/sound_datasets/marco")
 
 def test_clip():
     default_clipid = "impulse_response/+90deg_011_OCT3D_2_FR"
-    dataset = marco.Dataset(TEST_DATA_HOME, version="default")
+    dataset = marco.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
 
     expected_attributes = {
@@ -34,7 +34,7 @@ def test_clip():
 
 def test_load_audio():
     default_clipid = "impulse_response/+90deg_011_OCT3D_2_FR"
-    dataset = marco.Dataset(TEST_DATA_HOME, version="default")
+    dataset = marco.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     audio_path = clip.audio_path
     audio, sr = marco.load_audio(audio_path)
@@ -47,14 +47,14 @@ def test_load_audio():
 def test_load_metadata():
     # dataset
     default_clipid = "impulse_response/+90deg_011_OCT3D_2_FR"
-    dataset = marco.Dataset(TEST_DATA_HOME, version="default")
+    dataset = marco.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     assert clip.microphone_info == ["OCT3D", "2", "FR"]
 
 
 def test_to_jams():
     default_clipid = "impulse_response/+90deg_011_OCT3D_2_FR"
-    dataset = marco.Dataset(TEST_DATA_HOME, version="default")
+    dataset = marco.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     jam = clip.to_jams()
 
