+++ soundata/datasets/dcase_birdVox20k.py
@@ -84,7 +84,7 @@ BIBTEX = """
 INDEXES = {
     "default": "1.0",
     "test": "sample",
-    "3.0": core.Index(
+    "1.0": core.Index(
         filename="dcase_birdVox20k_index_1.0.json",
         url="https://zenodo.org/records/11176775/files/dcase_birdVox20k_index_1.0.json?download=1",
         checksum="d68016f669df15b67b5af1c4043593b9",
