+++ soundata/datasets/tau2020sse_nigens.py
@@ -80,9 +80,9 @@ BIBTEX = """
 """
 
 INDEXES = {
-    "default": "1.2",
+    "default": "1.2.0",
     "test": "sample",
-    "1.2": core.Index(
+    "1.2.0": core.Index(
         filename="tau2020sse_nigens_index_1.2.0.json",
         url="https://zenodo.org/records/11176861/files/tau2020sse_nigens_index_1.2.0.json?download=1",
         checksum="3ec732e50263d352070a32bc4af73eb1",
@@ -144,12 +144,7 @@ REMOTES = {
         filename="metadata_eval.zip",
         url="http://zenodo.org/record/4064792/files/metadata_eval.zip?download=1",
         checksum="f3584166d9a63b43c1e301b6fb722293",
-    ),
-    "index": download_utils.RemoteFileMetadata(
-        filename="tau2020sse_nigens_index_1.2.0.json",
-        url="https://zenodo.org/records/11176861/files/tau2020sse_nigens_index_1.2.0.json?download=1",
-        checksum="3ec732e50263d352070a32bc4af73eb1",
-    ),
+    )
 }
 
 LICENSE_INFO = """
