+++ soundata/datasets/dcase23_task6a.py
@@ -86,9 +86,9 @@ INDEXES = {
     "1.0": core.Index(
         filename="dcase23_task6a_index_1.0.json",
         url="https://zenodo.org/records/11176789/files/dcase23_task6a_index_1.0.json?download=1",
-        checksum="67415822f11bdb4400e422f0136a22f8"
+        checksum="67415822f11bdb4400e422f0136a22f8",
     ),
-    "sample": core.Index(filename="dcase23_task6a_index_1.0_sample.json")
+    "sample": core.Index(filename="dcase23_task6a_index_1.0_sample.json"),
 }
 
 REMOTES = {
@@ -146,7 +146,7 @@ REMOTES = {
         filename="clotho_metadata_test.csv",
         url="https://zenodo.org/records/3865658/files/clotho_metadata_test.csv?download=1",
         checksum="52f8ad01c229a310a0ff8043df480e21",
-    )
+    ),
 }
 
 LICENSE_INFO = """
