+++ soundata/datasets/dcase23_task6b.py
@@ -89,7 +89,7 @@ INDEXES = {
         url="https://zenodo.org/records/11176793/files/dcase23_task6b_index_1.0.json?download=1",
         checksum="66def2c298050d30ad9661d3e824c6b0",
     ),
-    "sample": core.Index(filename="dcase23_task6b_index_1.0_sample.json")
+    "sample": core.Index(filename="dcase23_task6b_index_1.0_sample.json"),
 }
 
 REMOTES = {
@@ -152,7 +152,7 @@ REMOTES = {
         filename="retrieval_captions.csv",
         url="https://zenodo.org/record/6590983/files/retrieval_captions.csv?download=1",
         checksum="f9e810118be00c64ea8cd7557816d4fe",
-    ) 
+    ),
 }
 
 LICENSE_INFO = """
