+++ soundata/datasets/singapura.py
@@ -135,7 +135,7 @@ INDEXES = {
         url="https://zenodo.org/records/11176844/files/singapura_index_1.0a.json?download=1",
         checksum="404d2057835cc97ef4dcef1b78e1a946",
     ),
-    "sample": core.Index(filename="singapura_index_1.0a_sample.json")
+    "sample": core.Index(filename="singapura_index_1.0a_sample.json"),
 }
 
 meta_files = [

[…]
