+++ setup.py
@@ -10,5 +10,4 @@ if __name__ == "__main__":
     setup(
         version=version_module.version,
         packages=find_packages(exclude=["test", "*.test", "*.test.*"]),
-        package_data={"soundata": ["datasets/indexes/*"]},
     )
