### 2024-05-17 #6634 @hadware
Everything seems good to me as well. I'd have liked that issue https://github.com/soundata/soundata/issues/159 and https://github.com/soundata/soundata/issues/160 to be addressed, but I understand it might be too much work to do in a short amount of time (and may also be out of the scope of this _review_). What do you think, @hagenw ?

Otherwise, i'm good, cheers to soundata's authors for their good handling of all of our comments :)
