### 2024-05-18 #6634 @hagenw
I would also highly recommend to work on those issues, but as I understand https://joss.readthedocs.io/en/latest/reviewer_guidelines.html, especially

> We like to think of JOSS as a ‘developer friendly’ journal. That is, if the submitting authors have followed best practices (have documentation, tests, continuous integration, and a license) then their review should be rapid.

I think, the current status matches already those best practices.
