### 2024-05-18 soundata/soundata#159 @faroit
Zenodo can be really slow from time to time. So if you rely on these files for unit tests I would advise against it.

Can't files be compressed and made into the free LFS limits?
