### 2024-05-20 #6634 @guillemcortes
Hello @hadware , @hagenw ,
Thanks for your useful comments that helped improving soundata! Regarding [soundata/#159](https://github.com/soundata/soundata/issues/159) I just want to mention that we are currently addressing it. You can check the progress here: https://github.com/soundata/soundata/pull/169. Our proposal is to host all indexes to Zenodo and only have "sample" indexes packaged with soundata (similar to tests/resources files). Regarding [soundata/#160](https://github.com/soundata/soundata/issues/160), we have it in mind and it is the next item in our todo list of next developments.
Thanks to both of you!
