### 2024-05-20 soundata/soundata#159 @guillemcortes
Hi @faroit , I've just mention it on the [soundata JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6634) but we think that a good compromise is to host the full indexes in zenodo and have a "sample indexes" for testing in the repo, similarly to the test audio and metadata files that we already have.
