+++ tests/datasets/test_esc50.py
@@ -13,7 +13,7 @@ TEST_DATA_HOME = os.path.normpath("tests/resources/sound_datasets/esc50")
 
 def test_clip():
     default_clipid = "1-104089-A-22"
-    dataset = esc50.Dataset(TEST_DATA_HOME, version="default")
+    dataset = esc50.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
 
     expected_attributes = {
@@ -40,7 +40,7 @@ def test_clip():
 
 
 def test_load_audio():
-    dataset = esc50.Dataset(TEST_DATA_HOME, version="default")
+    dataset = esc50.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip("1-104089-A-22")
     audio_path = clip.audio_path
     audio, sr = esc50.load_audio(audio_path)
@@ -53,7 +53,7 @@ def test_load_audio():
 def test_to_jams():
     # Note: original file is 5 sec, but for testing we've trimmed it to 1 sec
     default_clipid = "1-104089-A-22"
-    dataset = esc50.Dataset(TEST_DATA_HOME, version="default")
+    dataset = esc50.Dataset(TEST_DATA_HOME, version="test")
     clip = dataset.clip(default_clipid)
     jam = clip.to_jams()
 
