+++ soundata/datasets/starss2022.py
@@ -431,14 +431,9 @@ class Dataset(core.Dataset):
     @core.cached_property
     def _metadata(self):
         # parsing the data from the filenames due to lack of metadata file
-        json_path = os.path.join(
-            os.path.dirname(os.path.realpath(__file__)),
-            "indexes/starss2022_index.json",
-        )
-
         metadata_index = {}
 
-        with open(json_path) as f:
+        with open(self.index_path) as f:
             starss2022_index = json.load(f)
             all_paths_filenames = list(starss2022_index["clips"].keys())
 
