+++ soundata/datasets/tau2021sse_nigens.py
@@ -428,14 +428,9 @@ class Dataset(core.Dataset):
     @core.cached_property
     def _metadata(self):
         # parsing the data from the filenames due to lack of metadata file
-        json_path = os.path.join(
-            os.path.dirname(os.path.realpath(__file__)),
-            "indexes/tau2021sse_nigens_index.json",
-        )
-
         metadata_index = {}
 
-        with open(json_path) as f:
+        with open(self.index_path) as f:
             taunigenssse2021_index = json.load(f)
             all_paths_filenames = list(taunigenssse2021_index["clips"].keys())
 
