+++ soundata/datasets/tau2019sse.py
@@ -408,14 +408,9 @@ class Dataset(core.Dataset):
     @core.cached_property
     def _metadata(self):
         # parsing the data from the filenames due to lack of metadata file
-        json_path = os.path.join(
-            os.path.dirname(os.path.realpath(__file__)),
-            "indexes/tau2019sse_index.json",
-        )
-
         metadata_index = {}
 
-        with open(json_path) as f:
+        with open(self.index_path) as f:
             tausse2019_index = json.load(f)
             all_paths_filenames = list(tausse2019_index["clips"].keys())
 
