### 2024-05-22 caf83f9
Revert dataset version change in test_core
