### 2024-05-23 e1ae599
addition to remove test files from pip
