+++ setup.py
@@ -9,5 +9,6 @@ version_module = version_sfl.load_module()
 if __name__ == "__main__":
     setup(
         version=version_module.version,
-        packages=find_packages(exclude=["test", "*.test", "*.test.*"]),
+                packages=find_packages(exclude=["test", "*.test", "*.test.*","tests", "tests.*", "*.tests.*", "tests/__pycache__", "tests/__pycache__/*"]),
+
     )
