### 2024-05-24 042615f
improving downloading logs and index documentation
