+++ docs/source/contributing.rst
@@ -284,6 +284,49 @@ See an example of how an index should look like:
     Note that in this examples we group ``audio_voice1`` and ``audio_voice2`` in a single clip because the annotation ``voice-f0`` annotation corresponds to their mixture. In contrast, the annotation ``voice-f0`` is extracted from the multiclip mix and it is stored in the ``multiclips`` group. The multiclip ``multiclip1`` has an additional clip ``multiclip1-mix.wav`` which may be the master clip, the final mix, the recording of ``multiclip1`` with another microphone.
 
 
+Providing index information in the loader
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Once the index is created, you should include it in the loader module. The index path should be indicated in the ``INDEXES`` variable in the loader module.
+The ``INDEXES`` variable is basically a dictionary indicating which index to load for each available version of a dataset.
+
+Let's visualize an example ``INDEXES`` for the ``urbansound8k`` loader:
+
+    .. code-block:: javascript
+
+        INDEXES = {
+            "default": "1.0",
+            "test": "sample",
+            "1.0": core.Index(
+                filename="urbansound8k_index_1.0.json",
+                url="https://zenodo.org/records/11176928/files/urbansound8k_index_1.0.json?download=1",
+                checksum="1c4940e08c1305c49b592f3d9c103e6f",
+            ),
+            "sample": core.Index(filename="urbansound8k_index_1.0_sample.json"),
+        }
+
+In this example, the ``INDEXES`` variable is a dictionary with the following keys:
+
+* ``default``: The default version of the dataset to be loaded in case no version is explicitly given in the ``initialize()`` method.
+* ``test``: In the key ``test``, we store a sample version of the dataset to be used for testing purposes.
+* ``1.0``: An available version of the dataset through the dataloader.
+* ``sample``: A sample version of the dataset to be used for testing purposes. See `testing indications <add_tests>`_. for more detail.
+In a nutshell, this is a one-clip version of the dataset index which is used to test the methods and classes of the new dataloader,
+while keeping the size of the repository as low as possible.
+
+The values of the dictionary are instances of the ``core.Index`` class. This class wraps up a downloader and validator for an index.
+As seen in the example, we have two ways to define an index:
+providing a URL to download the index file, or by providing the filename of the index file, assuming it available locally.
+
+* The full indexes for each version of the dataset should be retrieved from our Zenodo community.
+See more details `here <upload_index_>`_.
+* The sample indexes should be locally stored in the ``tests/indexes/`` folder, and directly accessed through filename.
+See more details `here <add_tests>`_.
+
+**Important:** We do recommend to set the highest version of the dataset as the default version in the ``INDEXES`` variable.
+However, if there is a reason for having a different version as the default, please do so.
+
+
 .. _upload_index:
 
 3. Uploading the index to an online repository
@@ -311,8 +354,9 @@ From a contributor point of view, you may create the index, store it locally, an
     .. note::
         *<dataset-id>* is the identifier we use to initialize the dataset using ``soundata.initialize()``. It's also the filename of your dataset module.
 
-Visibility should be set as *Public*. There is no need to fill up anything else. 
+Visibility should be set as *Public*. There is no need to fill up anything else.
 All the information that users may need is found in the dataloader and the corresponding documentation.
+
 
 
 .. _create_module:
@@ -327,6 +371,8 @@ To quickstart a new module:
 2. Find & Replace ``Example`` with the <your_dataset_name>.
 3. Remove any lines beginning with `# --` which are there as guidelines. 
 
+You should follow the provided template as much as possible, and use the recommended functions and classes.
+
[…]
