### 2024-05-28 dec3b0c
fix list versions test
