### 2024-05-28 02d2490
minor doc improvements
