+++ docs/source/contributing.rst
@@ -284,75 +284,6 @@ See an example of how an index should look like:
     Note that in this examples we group ``audio_voice1`` and ``audio_voice2`` in a single clip because the annotation ``voice-f0`` annotation corresponds to their mixture. In contrast, the annotation ``voice-f0`` is extracted from the multiclip mix and it is stored in the ``multiclips`` group. The multiclip ``multiclip1`` has an additional clip ``multiclip1-mix.wav`` which may be the master clip, the final mix, the recording of ``multiclip1`` with another microphone.
 
 
-Providing index information in the loader
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-Once the index is created, you should include it in the loader module. The index path should be indicated in the ``INDEXES`` variable in the loader module.
-The ``INDEXES`` variable is basically a dictionary indicating which index to load for each available version of a dataset.
-
-Let's visualize an example ``INDEXES`` for the ``urbansound8k`` loader:
-
-    .. code-block:: javascript
-
-        INDEXES = {
-            "default": "1.0",
-            "test": "sample",
-            "1.0": core.Index(
-                filename="urbansound8k_index_1.0.json",
-                url="https://zenodo.org/records/11176928/files/urbansound8k_index_1.0.json?download=1",
-                checksum="1c4940e08c1305c49b592f3d9c103e6f",
-            ),
-            "sample": core.Index(filename="urbansound8k_index_1.0_sample.json"),
-        }
-
-In this example, the ``INDEXES`` variable is a dictionary with the following keys:
-
-* ``default``: The default version of the dataset to be loaded in case no version is explicitly given in the ``initialize()`` method.
-* ``test``: In the key ``test``, we store a sample version of the dataset to be used for testing purposes.
-* ``1.0``: An available version of the dataset through the dataloader.
-* ``sample``: A sample version of the dataset to be used for testing purposes. See `testing indications <add_tests>`_. for more detail.
-In a nutshell, this is a one-clip version of the dataset index which is used to test the methods and classes of the new dataloader,
-while keeping the size of the repository as low as possible.
-
-The values of the dictionary are instances of the ``core.Index`` class. This class wraps up a downloader and validator for an index.
-As seen in the example, we have two ways to define an index:
-providing a URL to download the index file, or by providing the filename of the index file, assuming it available locally.
-
-* The full indexes for each version of the dataset should be retrieved from our Zenodo community. See more details `here <upload_index_>`_.
-* The sample indexes should be locally stored in the ``tests/indexes/`` folder, and directly accessed through filename. See more details `here <add_tests>`_.
-
-**Important:** We do recommend to set the highest version of the dataset as the default version in the ``INDEXES`` variable.
-However, if there is a reason for having a different version as the default, please do so.
-
-
-.. _upload_index:
-
-3. Uploading the index to an online repository
-----------------------------------------------
-
-We store all dataset indexes in an online repository on Zenodo.
-To use a dataloader, users may retrieve the index running the ``dataset.download()`` function that is also used to download the dataset.
-To download only the index, you may run ``.download(["index"])``. The index will be automatically downloaded and stored in the expected folder in Soundata.
-
-From a contributor point of view, you may create the index, store it locally, and develop the dataloader.
-All JSON files in ``soundata/indexes/`` are included in the .gitignore file, 
-therefore there is no need to remove it when pushing to the remote branch during development, since it will be ignored by git.
-
[…]
