+++ tests/test_core.py
@@ -194,7 +194,7 @@ def test_list_versions():
         == "Available versions for urbansound8k: ['1.0']. Default version: 1.0"
     )
     with pytest.raises(ValueError):
-        soundata.list_dataset_versions("asdf") 
+        soundata.list_dataset_versions("asdf")
 
 
 def test_dataset_versions():
