### 2024-05-29 c57c736
Improve installing soundata for dev documentation
