+++ docs/source/contributing.rst
@@ -25,7 +25,7 @@ Installing soundata for development purposes
 
 To install Soundata for development purposes:
 
-    - First run: ``git clone https://github.com/soundata/soundata.git``
+    - First, run ``git clone https://github.com/soundata/soundata.git``
 
     - Then, after opening source data library you have to install all the dependencies:
 
