### 2024-05-29 48aff90
add warning to add index checksum before running test_full_dataset
