### 2024-05-30 soundata/soundata#156 @guillemcortes
Closing this since it has been addressed in #169
