### 2024-05-31 #6634 @guillemcortes
Hello @faroit,

We've merged https://github.com/soundata/soundata/pull/169, which fixed a couple of issues raised by @hadware and @hagenw. As I understand it, we have complied with the checklists but we don't know what the next steps are. So please let us know if you need anything else from our side to move forward with the review.
