### 2024-06-04 #6634 @faroit
Hi @guillemcortes thanks a lot for your work on the package. It looks like we can indeed proceed with the next steps. Can you draft a new release of `sounddata`?
