+++ .github/workflows/formatting.yml
@@ -12,7 +12,7 @@ jobs:
   black:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: black on soundata
         uses: psf/black@stable
         with:
