### 2024-06-10 #6634 @guillemcortes
Hi @faroit !

We've made the last updates to the paper and now it is ready to be published. (I'll generate the pdf below)
Zenodo record DOI of the software release v1.0.0: https://doi.org/10.5281/zenodo.11536559
The release can also be found on [github](https://github.com/soundata/soundata/releases/tag/v1.0.0) and [pypi](https://pypi.org/project/soundata/).

Let me know if there's something else that we have to do on our end, thanks!

Checklist:
