### 2024-06-10 9d33748
move import display plot utils to method, update docs
