### 2024-06-10 ccf27f5
test for new try/except
