+++ tests/test_core.py
@@ -323,7 +323,7 @@ def test_dataset_errors():
     #     d._clipgroup("a")
 
 
-def test_import_failure():
+def test_missing_plots_deps():
     # Simulate missing module by temporarily removing it from sys.modules
     with patch.dict("sys.modules", soundata=None):
         with pytest.raises(expected_exception=SystemExit):
