### 2024-06-11 9bfe066
Move import display_port_utils into explore_dataset() (#178)
* move import display plot utils to method, update docs

* test for new try/except

* improved test name

* ignore explore_dataset from tests

* formatting

---------

Co-authored-by: Guillem Cortès-Sebastià <cortes.sebastia@gmail.com>
