+++ tests/test_core.py
@@ -1,5 +1,6 @@
 import pytest
 import os
+import sys
 import numpy as np
 
 import soundata
