+++ soundata/version.py
@@ -2,4 +2,4 @@
 """Version info"""
 
 short_version = "1.0"
-version = "1.0.0"
+version = "1.0.1"
