### 2024-06-11 #6634 @genisplaja
Hello @faroit. We've released a new version of Soundata today fixing a small bug happening in the display utils for some Mac M1 users. Let us provide here the correct DOI: [10.5281/zenodo.11580085](https://doi.org/10.5281/zenodo.11580085).

Thanks and sorry for the inconvenience! Please let us know if there is something else to do from our side.
