### 2024-06-17 #6634 @danielskatz
👋 @magdalenafuentes -

Another editor suggests that you

> seem to use the citations syntax wrong , e.g.
```...with modelling and evaluation libraries ([@tensorflow, @pedregosa2011scikit, @mesaros2016metrics](#)), open sourcing models ([@zinemanas2020dcase, @speechbrain](#)) and data dissemination ...```
should be
```...with modelling and evaluation libraries [@tensorflow; @pedregosa2011scikit; @mesaros2016metrics], open sourcing models [@zinemanas2020dcase; @speechbrain] and data dissemination ...```
and single paper citations such as ```([@bittner2019mirdata](#))``` should simply be ```[@bittner2019mirdata]```.

Please take a look at the JOSS [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and make changes to match, then we can try the process again
