+++ paper.md
@@ -54,11 +54,11 @@ bibliography: paper.bib
 
 # Statement of need
 
-As research pipelines become increasingly complex, it is key that their different components are reproducible. In recent years, the research community has made considerable efforts towards standardization and reproducibility, with modelling and evaluation libraries ([@tensorflow; @pedregosa2011scikit; @mesaros2016metrics]), open sourcing models ([@zinemanas2020dcase; @speechbrain]) and data dissemination using resources such as [Zenodo](https://zenodo.org). However, discrepancies in the local version of the data and different practices in loading and parsing datasets can lead to considerable differences in performance results, which is misleading when comparing methods ([@bittner2019mirdata]). Besides, it is extremely inefficient to develop pipelines from scratch for loading and parsing a dataset for each researcher or team each time, and this increases the chances of bugs and differences that hinder reproducibility.
+As research pipelines become increasingly complex, it is key that their different components are reproducible. In recent years, the research community has made considerable efforts towards standardization and reproducibility, with modelling and evaluation libraries [@tensorflow; @pedregosa2011scikit; @mesaros2016metrics], open sourcing models [@zinemanas2020dcase; @speechbrain] and data dissemination using resources such as [Zenodo](https://zenodo.org). However, discrepancies in the local version of the data and different practices in loading and parsing datasets can lead to considerable differences in performance results, which is misleading when comparing methods [@bittner2019mirdata]. Besides, it is extremely inefficient to develop pipelines from scratch for loading and parsing a dataset for each researcher or team each time, and this increases the chances of bugs and differences that hinder reproducibility.
 
-`Soundata` is based on and inspired by `Mirdata` ([@bittner2019mirdata]), the popular library for working with Music Information Research (MIR) datasets, which share goals and vision. However, in MIR, the aforementioned issues are exacerbated due to the intrinsic commercial nature of music data, since it is very difficult to get licenses to distribute music recordings openly. Since musical datasets are extremely complex compared to other audio datasets, using the same software package for handling music and other audio datasets would lead to a very complex, hard-to-manage repository, which would be difficult to scale. Instead, we introduce `Soundata` as a separate effort that specifically addresses the annotation types and formats required by communities like DCASE[^dcase], which work with bioacoustics, environmental, urban, and spatial sound datasets.
+`Soundata` is based on and inspired by `Mirdata` [@bittner2019mirdata], the popular library for working with Music Information Research (MIR) datasets, which share goals and vision. However, in MIR, the aforementioned issues are exacerbated due to the intrinsic commercial nature of music data, since it is very difficult to get licenses to distribute music recordings openly. Since musical datasets are extremely complex compared to other audio datasets, using the same software package for handling music and other audio datasets would lead to a very complex, hard-to-manage repository, which would be difficult to scale. Instead, we introduce `Soundata` as a separate effort that specifically addresses the annotation types and formats required by communities like DCASE[^dcase], which work with bioacoustics, environmental, urban, and spatial sound datasets.
 
-There are other libraries that handle datasets like `Tensorflow` ([@tensorflow]) or `Tensorflow Datasets` ([@tensorflow_datasets]), `DCASE-models` ([@zinemanas2020dcase]) or `HuggingFace Datasets` ([@lhoest2021datasets]). But none of them serves as a stand-alone library that can easily be plugged into different work pipelines, with different modeling software.  Having a community-centric, open-source, audio-specialized library allows us greater flexibility to incorporate more audio-specific API functionalities and align our priorities with those of the audio community.
[…]
