+++ paper.bib
@@ -40,7 +40,8 @@ url={https://doi.org/10.48550/arXiv.1605.08695}
 }
 
 @misc{tensorflow_datasets,
-  title = {{TensorFlow Datasets}, A collection of ready-to-use datasets},
+  title = {{TensorFlow Datasets:} A collection of ready-to-use datasets},
+  author = {TensorFlow},
   howpublished = {\url{https://www.tensorflow.org/datasets}},
   year  = {2019}
 }
