### 2024-06-17 1832ef9
minor changes in paper.md
