+++ paper.bib
@@ -47,7 +47,7 @@ url={https://doi.org/10.48550/arXiv.1605.08695}
 }
 
 @inproceedings{zinemanas2020dcase,
-  title={{DCASE-models:} a phyton library for computational environmental sound analysis using deep-learning models},
+  title={{DCASE-models:} a {P}ython library for computational environmental sound analysis using deep-learning models},
   author={Zinemanas, Pablo and Hounie, Ignacio and Cancela, Pablo and Font Corbera, Frederic and Rocamora, Mart{\'\i}n and Serra, Xavier},
   booktitle={Proceedings of the Fifth Workshop on Detection and Classification of Acoustic Scenes and Events (DCASE 2020)},
   editor={Nobutaka Ono and Noboru Harada and Yohei Kawaguchi and Annamaria Mesaros and Keisuke Imoto and Yuma Koizumi and Tatsuya Komatsu},
