### 2024-06-17 #6634 @guillemcortes
Seen.  will merge now.

One more thing, I was actually thinking now about the following paragraph:

> Increase reproducibility: Provides a common framework for researchers to compare and validate their data. It also allows researchers to easily propagate dataset updates or fixes to the audio community, ensuring that methods are still comparable and researchers have the same up-to-date dataset versions. On that note, Soundata is designed to handle multiple versions of the same dataset, allowing transparent access to all versions of the dataset.

I would change "researchers" to "users" (to match the rest of the article) and also maybe substituting the second one by "them". What do you think? It will look like this:

> Increase reproducibility: Provides a common framework for users to compare and validate their data. It also allows them to easily propagate dataset updates or fixes to the audio community, ensuring that methods are still comparable and users have the same up-to-date dataset versions. On that note, Soundata is designed to handle multiple versions of the same dataset, allowing transparent access to all versions of the dataset.
