### 2024-06-17 #6634 @guillemcortes
Merged, can you regenerate the final proof pdf again, please?

Thanks!
