### 2024-06-18 #6634 @danielskatz
Congratulations to @magdalenafuentes ([Magdalena Fuentes]), @guillemcortes, and co-authors on your publication!!

JOSS depends on volunteers, and we couldn't do this without you
