### 2024-07-03 #6634 @faroit
👋 @genisplaja @guillemcortes @magdalenafuentes @hagenw @hadware

I just wanted to add here that I'm super happy how this submission came out. We not just had a rather quick turnaround time but also the submission substantially improved with the help of the very detailed and timely reviews as well as how they were addressed. After all, the communication on all sides was very effective and as such, a very positive example on how JOSS is designed to be.

If we would have a best-paper/best-reviewer award, I would suggest this submission for both! Thanks a lot, everyone!
