### 2024-07-03 #6634 @magdalenafuentes
👋 @faroit and all,

The submission experience was great for us too, and we're immensely grateful to you, and to @hagenw and @hadware for the great reviewing work and suggestions. Soundata will be featured in [JOSSCast](https://blog.joss.theoj.org/2024/01/introducing-josscast) and we'll make sure to talk about this very positive experience 🙏
