### 2024-07-05 #6634 @guillemcortes
👋 everyone,

Thank you for the kind words, @faroit. I second @magdalenafuentes words. It has been a very enriching experience and we only have good words for all of you. Thanks again!
